Time: 7 ms (83.62%) | Memory: 8.9 MB (83.40%) - LeetSync

+++ 3-longest-substring-without-repeating-characters/longest-substring-without-repeating-characters.cpp
@@ -1,16 +1,17 @@
 class Solution {
 public:
     int lengthOfLongestSubstring(string s) {
+        // s = "dvdf"
         int longest = 0;
         string holder;
 
-        for (int i = 0; i < s.size(); i++) {
-            int pos = holder.find(s[i]);
+        for (int i=0;i<s.size();i++){
+            size_t pos = holder.find(s[i]);
             if (pos == -1) {
                 holder += s[i];
                 longest = max(longest, static_cast<int>(holder.size()));
-            } else {
-                holder.erase(0, pos + 1);
+            }else {
+                holder.erase(0, pos+1);
                 holder += s[i];
             }
         }
